Add a note on putting the site in a subdir
I'm not sure this is necessarily the best way to show people how to make their resulting twitter site work in a subdir. But I suspect a lot of people will _want_ to do this, so having instructions available early on as part of setup is good.
I'm not suggesting using `--pathprefix` because `npm run build --pathprefix=/whatever/` doesn't work, because the script definition in package.json doesn't pass on extra parameters, and also because then you have to remember to add `--pathprefix` every time you rebuild, which seems a bit unfriendly.
Perhaps `eleventy.config.js` ought to explicitly return `{pathPrefix: "/"}` with a comment above it explaining that this is what you need to change for subdir support, and then the readme can simply say "if you want a subdir, edit `eleventy.config.js`"? Not sure what's best, so this PR is more a proposal to start discussion than a definitive thing!